+++ GameLogic/src/main/java/com/grupo04/gamelogic/gameobjects/Spawner.java
@@ -6,14 +6,11 @@ import com.grupo04.engine.Graphics;
 import com.grupo04.engine.Vector;
 import com.grupo04.gamelogic.BallColors;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class Spawner extends GameObject {
-    int totalWidth;
+    float totalWidth;
     int numRows;
-    int bubbleRadius;
-    int offset;
+    float bubbleRadius;
+    float offset;
     float firstHeight;
     int numBubbles;
     int bubblesPerRow;
@@ -24,7 +21,7 @@ public class Spawner extends GameObject {
 
     // width = ancho del conjunto de bolas, rows = "alto" del conjunto de bolas,
     // r = radio de las bolas, wallThickness = offset a partir del cual renderiza
-    public Spawner(int width, int rows, int r, int wallThickness) {
+    public Spawner(float width, int rows, float r, float wallThickness) {
         super();
         this.totalWidth = width;
         this.numRows = rows;
@@ -34,7 +31,7 @@ public class Spawner extends GameObject {
 
         // Calculamos el numero total de bolas que caben
         this.numBubbles = 0;
-        this.bubblesPerRow = this.totalWidth / (this.bubbleRadius * 2);
+        this.bubblesPerRow = (int)(this.totalWidth / (this.bubbleRadius * 2));
         this.bubbles = new Color[this.numRows][this.bubblesPerRow];
         for (int row = 1; row <= this.numRows; row++) {
             // Para filas pares, hay una bola menos
@@ -54,7 +51,7 @@ public class Spawner extends GameObject {
             for (int j = 0; j < bPerRow; ++j) {
                 graphics.setColor(this.bubbles[row-1][j]);
                 graphics.fillCircle(new Vector(this.offset + ((row % 2 == 0) ? this.bubbleRadius : 0) + this.bubbleRadius + this.bubbleRadius * 2 * j,
-                        this.offset + (33) /*Cuando se solucione lo del redimensionado, quitarlo*/ + firstHeight  + firstHeight * 2 * (row - 1)), this.bubbleRadius);
+                        this.offset + (33)/*?*/ + firstHeight  + firstHeight * 2 * (row - 1)), this.bubbleRadius);
             }
         }
     }
